refactor(bezier-codemod): separate utils directory and shared directory
