+++ packages/bezier-codemod/src/transforms/v2-typography-interpolation-to-css-variable/transform.ts
@@ -5,7 +5,7 @@ import {
 } from 'ts-morph'
 
 import { removeUnusedNamedImport } from '../../utils/import.js'
-import { interpolationTransform } from '../../utils/interpolation.js'
+import { interpolationTransform } from '../../shared/interpolation.js'
 
 const WARNING_COMMENT = '/* NOTE: Do not use font-related css variables below separately, use Text component instead */'
 const getFontSize = (num: number) => `font-size: var(--typography-size-${num}-font-size);`
